requirements.txt
boto3>=1.26.0
pandas>=1.5.0
numpy>=1.21.0
streamlit>=1.22.0
plotly>=5.13.0
faker>=18.0.0
pytest>=7.3.1
sagemaker>=2.150.0
pyspark>=3.4.0
aws-glue-sessions>=1.0.0 